Reduce code duplication with javasrc interop
