+++ joern-cli/frontends/javasrc2cpg/src/main/scala/io/joern/javasrc2cpg/typesolvers/EagerSourceTypeSolver.scala
@@ -4,7 +4,7 @@ import com.github.javaparser.ast.body.TypeDeclaration
 import com.github.javaparser.resolution.declarations.ResolvedReferenceTypeDeclaration
 import com.github.javaparser.symbolsolver.javaparsermodel.JavaParserFacade
 import com.github.javaparser.symbolsolver.model.resolution.{SymbolReference, TypeSolver}
-import io.joern.javasrc2cpg.JpAstWithMeta
+import io.joern.javasrc2cpg.passes.JpAstWithMeta
 import org.slf4j.LoggerFactory
 
 import scala.jdk.CollectionConverters._
